refactor Gameover; add hours played if >60mins
I saw someone share a screencap of ScreenGameover because he was proud of the total time he'd played that session – "202min. 53sec."  I paused for a moment to mentally calculate how many hours that was and immediate understood that the UI could be improved.  I didn't anticipate people regularly playing for >60 minutes!

Thus, this commit adds a calculation for hours to ScreenGameover.  It doesn't appear if the session was < 60 minutes.

Additionally, this commit somewhat refactors ScreenGameover to have slightly less duplicated code.