show time elapsed in SSM header in EventMode
Several people requested a quick way of gauging how long they'd been playing in EventMode as part of their workout/exercise routine.  This would not have occurred to me since I never play with EventMode turned on, but I can see its usefulness.

This commit adds a timer that counts up into the header of ScreenSelectMusic if EventMode is enabled.  The timer starts when ScreenSelectMusic is loaded for the first time (at the start of a new game session), and resets when the SL table is reset (ScreenTitleJoin).

It is a general timer in that it includes time spent picking a song on ScreenSelectMusic, choosing mods in ScreenPlayerOptions, playing in ScreenGameplay, and basking in success/failure on ScreenEvaluation.  This is in contrast to the time that appears on ScreenGameover, which only factors in time spent in Gameplay.

This feature could be made optional (a ThemePref) in a future commit depending on feedback.