Fix scatter plot bug
This is necessary because CurrentSecond includes the offset of whenever the player hit the note, so if you hit the last note(s) super late or the first note(s) super early it will be outside the ends of the song and fall off the graph.

But for some reason even after applying this, it STILL falls off the right end (and only the right end). So now there's padding. Maybe the JudgementMessage command takes a few frames to propagate? or GameState precision isn't the best?