attempt to catch NoteSkin Lua errors via pcall()
This adds a pcall() that attempts to load a NoteSkin actor before actually adding it to the main ActorFrame on ScreenPlayerOptions overlay.  I don't fully understand how to use pcall() yet, but this code seems to accomplish the desired result.

That is, NoteSkins with Lua errors don't get previews but also don't halt the rest of the Lua in ScreenPlayerOptions overlay.lua, preventing players from choosing a speed mod.  Enough people reported this issue (on Reddit, on ZIv, on GitHub, on Facebook, etc.) that this was worth addressing, if possible.

I'll note that EVERY NoteSkin I have throws an error of "attempt to call a table value".  This suggests I'm doing something wrong, but I haven't had time to dig into it yet.