improve pcall usage in ScreenPlayerOptions
Previously, we were reading every NoteSkin without Lua errors from disk twice, once to check for errors, and once to add it to the main ActorFrame.

This commit changes my usage of pcall so that it is only necessary to read each from disk once.  This is possible by making use of pcall's second return value, which will either be the NoteSkin actor that was successfully loaded or nil.