allow StepStats when (Center1Player && WideScreen)
Multiple users posted in various places (Facebook, Reddit, etc.) asking how to enable StepStatistics, not realizing that it wasn't an available option when Center1Player was enabled.

In truth, it was previously not available under those conditions because I didn't want to do the work of resizing and repositioning the panel.  This commit resolves that.  It is not possible to see live StepStatistics while Center1Player is enabled IF you are using WideScreen (16:9 and 16:10 are both supported).

Visually, it's a bit cramped, and the code here is rather kludgy, but it works.

With the Rhythm Horizon public beta less than a month away this will hopefully be the final commit to Simply Love for SM5.