don't use _fallback graphics for ScreenSetBGFit
This commit attempts to improve the usability of ScreenSetBGFit in Simply Love.  It previously used the _fallback theme's code and assets.

I've pulled in ScreenSetBGFit overlay.lua from _fallback and slightly modified it.  Thankfully, the implementation as it existed in _fallback was heavily commented and easy enough to follow along with.  Thanks, kyzentun.

I've also replaced the proof-of-concept/placeholder checkerboard png files with assets that look more like background files for songs.  (Spoilers: I yoinked the bg files from the stock "StepMania 5" song pack.)