fix EvalSummary to support absence of menubuttons
It seems that with my previous code, it was only possible to navigate through multiple pages of scores in EvaluationSummary using menu buttons.  This potentially left home-pad players in a pickle.

This commit checks the OnlyDedicatedMenuButtons preference, assuming that if it is disabled, the player wants to be able to navigate this screen using gameplay buttons.  If OnlyDedicatedMenuButtons is enabled, the screen should function the same.

This also adds support for navigating this screen using DownLeft and DownRight on a 5-panel/pump pad.