make SL-PlayerOptions translatable
As the inline comments state:

Each OptionRow needs choices to present to the players.  Sometimes these choices can be hardcoded strings, but many times it will be nice if they can be translatable.

For each of these subtables, you can specify a "Choices" function or a "Values" function.

Both are defined here as a table of strings, but a Values table will require corresponding language strings in en.ini (or es.ini, fr.ini, etc.) whereas the strings in a Choices table will presented to the players as-is.

Here's hoping I didn't break anything along the way?