add "Exit" code for ScreenSeletCasual
This commit adds the [MenuLeft, MenuLeft, MenuRight, MenuRight, MenuLeft, MenuLeft, MenuRight, MenuRight] code to ScreenSelectCasual.

If EventMode is on, it will redirect the players to either the title screen or ScreenEvaluationSummary depending on whether songs were played.

If EventMode is off, the code will redirect players to ScreenSelectMusic in Competitive gamemode.  I have often observed people playing on a public cabinet choose the wrong gamemode (Casual when they meant to choose Competitive), so this helps that.

This commit also cleans up the header for ScreenSelectMusicCasual and removes some unused variables.