SSMCasual - fewer BMT actors
This refactors SSMCasual's to use few BitmapText actors.  Previously, each song item in the wheel had a unique BMT for title, artist, length, bpm, and genre that were hidden/displayed as needed.

Now there is a single set of BMTs and the text is updated as needed.

By way this simplification, I no longer have to figure out how to handle the BMT xy tweens when players were scrolling rapidly...