SSMCasual - cover flow for group selection
SSMCasual now uses cover flow navigation for group selection as opposed to the former 3x3 grid.  This feels better, but not done yet.

I had to slightly modify ./Scripts/Consensual-sick_wheel.lua by adding an optional sixth argument to the create_actors method, a boolean for SetDrawByZPosition().  I needed to be able to set it to true for the cover flow effect here.

In an effort to not disrupt the other screens already relying on sick_wheel, the argument is optional and will default to false if not passed in (every other screen that uses sick_wheel except SSMCasual now).