allow DefaultSong(s) for SSMCasual
Machine operators can now specify one or more default songs for ScreenSelectMusicCasual.  If one song is specified, SSMCasual will default to that when for the first round.  If multiple valid songs are specified, one will be randomly selected for each new casual mode game session.

I'm going to ship the next release with CasualMode-DefaultSong.txt already filled in with some songs.  If none of those songs exist on a given machine, SSMCasual will default to the first song of the first pack found in ./Other/CasualMode-Groups.txt

If ./Other/CasualMode-Groups.txt is empty or contains only invalid groups, SSMCasual will default to the first song from the first valid pack found on the machine.