better handling of MenuTimer on SSMCasual
The previous commit mostly fixed an issue with forcing mods on Players if they didn't actively select a song before the MenuTimer expired.

This improves on it by respecting any mods the players may have chosen in the event that they had, say, chosen a difficulty, but not quite made it to confirming the song choice before time ran out.