better group pruning for SSMCasual
I tried testing SSMCasual in techno and pump, and found the screen broken in techno.  My group pruning functions were:

1. finding strings in ./Other/CasualMode-Groups.txt
2. pruning them all out because none had techno steps
3. returning empty tables and calling it a day

I now check to see if the pruning function returns an empty table and, if so, trying pruning again using the full list of all groups available to StepMania.  This is more likely to find something valid (though still not guaranteed).

While this is likely a very uncommon scenario, it's better to be fixed.

I should probably followup with some fix for the case where no valid groups are found even with this fix.  I think a message letting players know what happened, instructing them to alert their machine operator, and allowing them to exit the screen is the best I can do there.