handle case of no songs available for SSMCasual
It's unlikely but possible for a machine to have no stepcharts that are valid for Casual mode.

If that happens, display a message informing the player of this, tell them to talk to their machine operator, and let them exit the game.  It'll be unfortunate for the player, but this is better than silently throwing Lua errors.