fix number of courses on Title Screen
It seems the integer that SONGMAN:GetNumCourses() returns always includes autogenerated group courses.  This led to confusion on the title screen with seemingly radically inflated numbers.  (On my personal machine, I have 7 courses with 13 trails total, but the Title Screen was saying I had "902 courses".)

This commit changes the behavior to correctly display the user's number of trails based on their AutogenGroupCourses preference.