improve ChartParser's chart-finding routine
The previous implementation made some string-matching assumptions when parsing .sm files that could, in edge cases, result in the ChartParser finding the wrong the chart within the .sm file.

Specifically, the Bearpocalypse 6 pack seems to tag charts with text that matches existing difficulty strings, but not the difficulty string for the chart in question.  For example, a chart with difficulty "Medium" might be tagged "Hard" and a chart with difficulty "Easy" might be tagged "Medium".

---

With this implementation, the ChartParser uses the idea that a single chart within a .sm file can be split into 7 tokens using : as the delimiter.  It then uses those 7 indexed tokens to assess whether the chart we've encountered in the .sm file is indeed the chart we're looking for.

This hopefully addresses issue #126 on GitHub.