do less work in TextBannerAfterSet
About a minute after making my previous commit which added comments explaining TextBannerAfterSet() it occurred to me that not all those Actor commands needed to be constantly reapplied every time a MusicWheelItem changed its text.

In fact, the only thing that really needed to change was the y position of the Title and Subtitle, depending on whether Subtitle was an empty string.  The values of the other commands (zoom, maxwidth, etc.) don't ever change, so they can be applied once in the appropriate OnCommands via Metrics.