fix more Credit-related bugs with SelectStyle
This fixes at least two more credit-related bugs in ScreenSelectStyle that were introduced in v4.8's rewrite of this screen.  The screen was rewritten as a Lua-driven ScreenWithMenuElements in v4.8 to sidestep negative credit bugs that could occur when the screen was managed by the engine.

1.  Previously, if the Premium preference was set to "Premium_Off", and both players were joined nearly simultaneously on the Title Screen, ScreenSelectStyle would initialize to only have "1 Player" as an immediately selectable choice.  In order to select "2 Players" or "Double", it was necessary to press START for whichever side the screen didn't consider to be a HumanPlayer yet...

This was pretty confusing/unclear, but the screen should initialize its choices properly now.

2. Previously, if the Premium preference was set to "Premium_DoubleFor1Credit" and both players were joined on the Title Screen (deducting 2 credits), the player would then end up losing 2 credits to play double, effectively nullifying the Premium setting. :(

This has been fixed by checking for this scenario and adding a credit back when it occurs.

Analogous situations with 1 Player costing 2 credits have been avoided by initializing the screen's Enabled choices properly as part of bug fix 1.  It should not be possible to join both players on the Title Screen and still pick 1 Player.