display overall pack lengths on SSM
This adds functionality to ScreenSelectMusic to display the overall duration of song packs when actively hovering over a group.  This was requested.

Prior to loading actors, pre-calculate each group's overall duration by looping through its songs and summing their durations.  Store each group's overall duration in a lookup table, keyed by group_name, to be retrieved + displayed when actively hovering on a group (not a song).

I haven't checked, but I assume that continually (re)calculating group durations on-the-fly would have performance ramifications when rapidly scrolling through the MusicWheel.

A consequence of pre-calculating and storing the group_durations like this is that live-reloading a song on ScreenSelectMusic via Control R might cause the group duration to then be inaccurate, until the screen is reloaded.

sent out into pitch
over the sea, past the sky,
please, please let it reach.