fix SelectStyle bug with Premium_DoubleFor1Credit
In 407cb459f0139449ccdd682cf95ae9e09d3d9394 I fixed some coin related bugs on ScreenSelectStyle involving Premium_Off and Premium_DoubleFor1Credit.  It was an improvement, but some bugs still slipped through the cracks.

This commit fixes a bug for Premium_DoubleFor1Credit.  If that's set, and we're on ScreenSelectStyle, and both players are already joined (2 credits already deducted correctly), the existing logic would have attempted to deduct a third credit if "versus" was chosen.  If there were no credits available to deduct, versus would appear as a valid choice but players wouldn't be able to actually pick it.

It's fixed now.

This also changes the behavior of the screen to automatically move the active choice if the current choice suddenly becomes invalid (single becomes invalid once both players are joined).  The old behavior wasn't resulting in errors, but this is a little more usable.

Thanks go to @karamel59 for identifying and offering fixes for these issues over at https://github.com/djpohly/Simply-Love-SM5/pull/1

Finally, this condenses GetNextEnabledLeft() and GetNextEnabledRight() into a single function, GetNextEnabledChoice().  It takes an integer argument – positive to move right, negative to move left.