better support for less common games
This commit introduces better support for kb7 and para; both should be playable in the sense that Lua errors shouldn't occur.

CHANGES:

• SelectStyle now supports para and kb7 by removing double as a choice for those

• Eval's Pane2 (per-column judgment breakdown) now works in para and kb7

• Eval's Pane5 (QR Code) now only appears in dance (that's all that GrooveStats supports)

• Moved QR Code help text to en.ini to support future localization

• ScreenTestInput (in the operator menu) only has the graphic assets to support dance, pump, and techno.  The screen is no longer accessible from other games.

• ScreenTestInput can now be exited by pressing Escape on a keyboard

• update en.ini to explain that kb7 and para are partially supported and that beat, kickbox, and lights are not playable in SL

beat and kickbox are different enough in their available styles that it probably isn't worthwhile to fix SL to support them.  lights isn't intended to be playable as far as I can tell.

I'd like to prevent players from ever starting a game in such unsupported games, but that will require me to rewrite the scroller in ScreenTitleMenu, so it will follow in a future commit.

---

In July, Van Gogh wrote that he had become absorbed "in the immense plain against the hills, boundless as the sea, delicate yellow."  He had first become captivated by the fields in May, when the wheat was young and green. In July he described to Theo "vast fields of wheat under turbulent skies."