small cleanup
Remove unnecessary ActorFrames.  Set visible(false) on ScreenGameplay in's "heart explode" actors once they're done animating and save a few clock cycles during Gameplay.

Abstract MeasureCounter and ModsLevel code out of ./ScreenGameplay in/default.lua and into a standalone file to better separate form from function.

---

Everything is more complicated than you think. You only see a tenth of what is true. There are a million little strings attached to every choice you make; you can destroy your life every time you choose. But maybe you won't know for twenty years. And you'll never ever trace it to its source. And you only get one chance to play it out. Just try and figure out your own divorce.

And they say there is no fate, but there is: it's what you create. Even though the world goes on for eons and eons, you are here for a fraction of a fraction of a second.

Most of your time is spent being dead or not yet born. But while alive, you wait in vain, wasting years, for a phone call or a letter or a look from someone or something to make it all right.

And it never comes or it seems to but doesn't really. And so you spend your time in vague regret or vaguer hope for something good to come along. Something to make you feel connected, to make you feel whole, to make you feel loved.