don't allow players to use unsupported games
This adds CurrentGameIsSupported() to SL-Helpers.lua.  It contains a list of games SL supports and returns true (supported) or nil (not supported).

At this time,
Supported games: dance, pump, techno, para, kb7
Unsupported games: (anything not explicitly in the above list)

ScreenInit has been updated to use this function and immediately redirect players to ScreenSelectGame if the current game is unsupported.  This handles

• changing the game in the operator menu
• starting StepMania with CurrentGame in Preferences.ini set to an unsupported game
• switching into SL from another theme in an unsupported game

It seems like it works, but probably merits further testing by someone else.

The other language files (es.ini, fr.ini, pt-br.ini) should definitely be updated to emphasize which games currently aren't supported.

---

I wanted to be a writer, that's all. I wanted to write about it all. Everything that happens in a moment. The way the flowers looked when you carried them in your arms. This towel, how it smells, how it feels, this thread. All our feelings, yours and mine. The history of it, who we once were. Everything in the world. Everything all mixed up, like it's all mixed up now. And I failed. I failed. No matter what you start with it ends up being so much less.