small improvements for ScreenEval in Casual mode
Diffuse the JudgmentLabels like we were doing in ECFA and StomperZ.  Color is nice.

Don't show the bpm and music rate Quad, and nudge the song title and banner down some since there is now more space.

Add a slight shadow to the help text (Press START To Continue) to make it more legible in RainbowMode.

---

It has also been suggested that the curse of knowledge could contribute to the difficulty of teaching.  It could be potentially ineffective, if not harmful, to think about how students are viewing and learning material by asking the perspective of the teacher as opposed to what has been verified by students.  The teacher already has the knowledge that he or she is trying to impart, but the way that knowledge is conveyed may not be the best for those who do not already possess the knowledge.