make TargetScoreGraph more available
The TargetScoreGraph code contributed by JackG a few years ago supported drawing both full-width and narrow-width graphs, but didn't always know when to draw which and didn't always place them correctly on the x-axis.  Since the file was unfamiliar and sizable, and I was much less comfortable with Lua and SM5 theming at that time, I opted to remove the OptionRow for TargetScoreGraph in various (typically duplicated and inconsistent) ways.

This commit cleans up a lot of the original logic for knowing when to draw which size graph and determing where to place them on the screen.  TargetScoreGraph is now available for all styles of all supported games.  You can now use it in everything from dance-double to techno-single8 to para-versus.  dance-solo even works.

I added many comments to the TargetScoreGraph code as I went and figured out how things were working.  Some functions that it exposed as global are now local to that file.

In general, the TargetScoreGraph code is still disorganized and quite difficult-to-follow, so Someone™ should restructure it some day.  Yeah.

I don't think it should be loaded on a per-player basis.  Players think of TargetScoreGraphs as a per-player option, but the file really isn't structured that way, and some optimizations can be made (fewer actors being added to the screen) if it is only loaded once.

There is currently a bug where having 2 narrow-width-graphs directly next to one another when the display is 4:3 will result in the ☆☆☆ text being cut off.  It could be more easily fixed if a single set of background Quads were drawn because the file were only loaded once.  I'm leaving it as an open bug for now; it can wait until the rewrite.

Internally, the theme now thinks of the "Data Visualizations" OptionRow as "DataVisualizations" instead of "TargetStatus".  TargetStatus no longer accurately reflects the nature of that OptionRow and I kept losing track of what I was doing until I changed it.

Existing player profiles will have a reference to TargetStatus that will be ignored the next time they are loaded and not written back upon next save.  It will go away of its own accord.

---

It opens with a proposition, "would you like to fly through the sky?"  Then, it starts moving.

Constantly moving forward, it carries us along.  We pass scenery and countryside and time, remembering quiet moments shared in the apartment.  The ceiling fan swings.

And it's beautiful, so beautiful, but painful too.

"Would you like to go together?"

ずっとそばにいるね。