allow ActionOnMissedTarget without Pacemaker
The ActionOnMissedTarget feature allows players in EventMode to have the song automatically restart or end if their target score becomes no longer achievable.  This was added to facilitate score grinding.

The logic for ActionOnMissedTarget depends on the logic for the Pacemaker text, but from an end user's perspective, the two features are entirely distinct.

Previously, the Pacemaker needed to be enabled for ActionOnMissedTarget to work which understandably confused many players.  It's now possible to have the Pacemaker or ActionOnMissedTarget or both or none.

In the SL table, "TargetBar" became "TargetScore" because that made more sense to me.

I quickly realized that the theme had already been using "TargetScore" to refer to the Pacemaker, so the old "TargetScore" became "Pacemaker".  This also made more sense to me.

I've updated the localization files to the best of my ability and added additional comments to TargetSCore/default.lua

---

If pressed to provide a label, I would use phrases like "contemporary fairytale" or "miserable allegory."  It's a tale about commitment, the complex systems that tacitly arise out of commitment, and the long-term effects of not acknowledging those systems.