update CasualMode-Groups.txt
Fraxtil and DJ OMiY continue to put out quality novice-friendly content.

---

Their secret marriage displeased the queen, and the two were forbidden to meet.  The girl weaver became the star Vega, the cowherd became the star Altair, and the Milky Way became the river that separated them.  But on the seventh day of the seventh month of the lunar calendar, the lovers cross a bridge of magpies to be together, if only for one day.