small changes/cleanup for MeasureCounter
Dim the MeasureCounter to be slightly grey when counting breaks; diffuse white when counting stream.  This helped me differentiate between the two a little more.

The break counter decrements to 0, but seeing it display 0 for a full measure before immediately changing to the stream counter seemed jarring (to me).  This changes the behavior to count down from x to 1, hide for a measure during "break 0" and then return for stream counting.

Update MeasureCounterPosition OptionRow to present "Left" and "Up" as choices and allow choosing one, the other, neither, or both.  The defaults (as defined in SL_Init.lua) are to have Left enabled and Up disabled to maintain the SL3.95 behavior as "standard."  Disable Left to center the measure counter.  Enable Up to nudge it up over the judgment graphic.

Update the language files to the best of my ability.

Deprecate MeasureCounterPosition in ./Scripts/SL-CustomProfiles.lua and replace it with MeasureCounterLeft and MeasureCounterUp flags.

Determining what counts as a break and what doesn't seems like it might benefit from a more robust+thoughtful heuristic, but that can wait until a future release.

---

The retrospect is glorious, the prospect is inspiring: Much might be said of both. But one idea dominates my mind. This — my best, my dearest — is for your noble cause.

I have observed electrical actions, which have appeared inexplicable. Faint and uncertain though they were, they have given me a deep conviction and foreknowledge, that ere long all human beings on this globe, as one, will turn their eyes to the firmament above, with feelings of love and reverence, thrilled by the glad news: "Brethren! We have a message from another world, unknown and remote. It reads: one... two... three..."