don't smooth offset data into next TimingWindow
Smoothing the offset histogram in Pane4 of ScreenEval is good overall, but high-level tech players have noted that their Quad Star histograms are wider than they should be.  The edges of a Quad Star histogram would sometimes be smoothed into W2 territory, causing concern.

Although the histogram was designed to help new players establish a sense of timing more quickly (i.e., it was not for high-level players consistently earning Quad Stars), this is a valid thing to point out.

With this commit, I'm leaving the smoothing in place, but adding some simple checks to ensure that the histogram doesn't get smoothed into a TimingWindow the player didn't actually earn.