there is nothing to suggest it
You run until your breathing sounds like crying, until you no longer recognize the sound of your body surviving.  For a moment you are something else, a reactionary assortment of bones and tissue, and there is relief in this.  You are the pain of your heart trying to regulate its beating to ensure the survival of the larger system.

As these things go, the panic in your body subsides quickly, and the panic in your brain resumes.  As these things go, your heart replaces one form of pain for the other.  You continue running.