Revert "the sound of your body continuing to survive"
This reverts commit f2de1b2b5e921bb5fb65a6011b6d363563f85fda.

It was a hopeful experiment, but players on arcade cabinets have already
reported that the 4.8.5 beta broke "the up+down shortcut to close open
groups" not realizing that (for whatever reason) MenuUp and MenuDown
were mapped in their Keymaps.ini, despite their arcade hardware not
having such physical buttons.

Reverting this for now.  It's no worse than before, just not any better,
either.