use math.log10() to find digits in note counts
In 5316b920dba209bee25c22aa40f9beb6072a3450, I added support for the StepStats side pane in Gameplay to accommodate more than 4 digits of notes.

If the total note count for a song is >= 9999, 4 digits will be used (and leftpadded with 0s as needed), and the display will grow to accommodate extra digits as needed.  For example:

 • "Last Trip" has 22897 notes and needs 5 digits
 • "140 BPM Stamina Killer" has over 100k notes and needs 6 digits
 • "1 week of 100 BPM Stream" has over 4M notes and needs 7 digits

The previous implementation used a while loop, division, and math.pow() to count digits.  This was probably fine since this calculation would only be performed once, prior to Init, but I'd also failed to remember that I needed the same calculation elsewhere in the same file, and was using a very sloppy (number -> string -> number) cast there.

This commit changes both of those to use the more standard math.log10() approach of determining digits in base-10.  It is currently unclear what performance impact this might have, or how I might even assess that.

I still need to handle positioning and zooming of the BitmapText when Center1Player is in effect, but that is a different matter.