Refactor density graph rendering (BG video supported!)
Previously for a scrolling NPS graph all the vertices were rendered
from the beginning and the screen edges and a "fake" background image
were used to hide the overhang.

After this only the vertices that should be visible are rendered. No
overlap truncation required and all background features are suppored
now: hiding, dimming, change, videos.

The calculations for positioning the NPS graph with Center1Player is a
bit messy, but I couldn't find a cleaner way to properly align it.