use bools in SL_Init Preferences where appropriate
Specifying a 0 or 1 in SL_Init.lua when the corresponding StepMania preference was defined in PrefsManager.h as <bool> would (I think) result in that 0 or 1 being stringified and then evaluated as truthy.

I can't quite follow along with the relevant C++, but various themeside experiments suggest that HarshHotLifePenalty, SubSortByNumSteps, and PercentageScoring were always being set to true in Simply Love with each new game cycle.

It's embarrassing I never realized this was an issue until now.  Thankfully, the actual impact was fairly minimal: HarshHotLifePenalty was always on in StomperZ when it should not have been.

This should resolve https://github.com/ddrillini/Simply-Love-SM5/issues/19

---

To review, if the StepMania preference is a boolean, the correct Lua syntax for PREFSMAN:SetPreference() is true/false, not 0/1.

I can see how past-me made this mistake.  In SM5's Preferences.ini, boolean Preferences are specified using 0 or 1.  For example, this is good Preferences.ini syntax:

# Preferences.ini
HarshHotLifePenalty=1
SubSortByNumSteps=0
PercentageScoring=1

while this is good Lua syntax:

-- some .lua file
PREFSMAN:SetPreference("HarshHotLifePenalty", true)
PREFSMAN:SetPreference("SubSortByNumSteps", false)
PREFSMAN:SetPreference("PercentageScoring", true)