user fewer masks in Gameplay
The Border() helper function in ./Scripts/SL-Helpers.lua uses three Quads and masking to quickly achieve a simple, reusable stroke effect.  This is normally fine, but sometimes resulted in visual oddities in Gameplay.

Some NoteSkins use masking to contain their scrolling textures within their 3D models, and that masking would interact with the masking Border() used in Gameplay's Standard LifeMeter and SongInfoBar.  This resulted in NoteSkins with masking drawing under the LifeMeter and SongInfoBar.  Since not all NoteSkins use masking, it was visually inconsistent.

My fix is to not use the Border() helper function in ScreenGameplay any longer, and to just use normal, run of the mill Quads.

People have been asking about this for years, but I am just getting around to fixing it now. :-/