don't require EventMode to see TestInput on Eval
Archi asked about this some months ago.

The original idea was that machine operators of public arcade cabinets wouldn't want random people attempting to diagnose/fix the pads.  Perhaps there are not enough public arcades running SM5 + SL to warrant this concern.

This pane does still require OnlyDedicatedMenuButtons to be enabled because without that, input would be eaten and players would be unable to navigate out of this pane once in.