prompt players before EscapeFromEventMode
Having a MenuButton code (MenuLeft,MenuLeft,MenuRight,MenuRight,MenuLeft,MenuLeft,MenuRight,MenuRight) for exiting a game session while on ScreenSelectMusic is handy for private cabinet owners who may not have a keyboard always connected.

Unfortunately, it's common to accidentally activate this input code while browsing the MusicWheel and have the session terminated early.

This commit adds a quick yes/no prompt ("Do you want to exit this game?") after the EscapeFromEventMode code is detected.

Note that, for now, this code is only available in EventMode but I'm going to think this over and perhaps change that.

There is still a similar input code in ScreenSelectMusicCasual that redirects players to ITG mode in case they've accidentally chosen Casual Mode while at their local anime convention, but that is handled separately.