fix update of BPM value text on SelectMusic
The BPM text in SelectMusic was not updating to be an empty string when scrolling over group items in the MusicWheel.  It was retaining the BPM value of the last-seen song.

This bug was introduced in f1dcf9a587b000ea4fccec4526729985fac14b1a.

It seems that GAMESTATE:GetCurrentSong() will return nil if the MusicWheel is on a group item, but that GAMESTATE:GetCurrentSteps(player) will return the last-seen steps object in the same circumstance.

This behavior seems inconsistent, but changing anything engine-side (especially ScreenSelectMusic) is fraught with complications and community fallout, so I'm fixing it here, theme-side.