use hardcoded numbers for NoteFieldWidth table
There really wasn't any advantage to defining each possible NoteField width as a function that would be calculated and recalculated each time it was called if it was only ever intended to return a single, unchanging number.

There's possibly an argument for making the NoteFieldWidth table global so that a themer could, for example, look up the NoteField width of pump routine while SM5 was currently in dance solo, but there's no realistic use for it in Simply Love at this time.

The next commit will update all calls to GetNotefieldWidth() throughout the theme.

---

-- chimerical (adjective)
-- chi·​me·​ri·​cal | \ kī-ˈmer-i-kəl
-- existing only as the product of unchecked imagination
-- solodouble = 768
-- soloversus = 768
--
-- #thetruthisoutthere
-- #iwanttobelieve