fix MenuTimer for SelectProfile
Previously, if the MenuTimer was enabled, it would reach 0 on ScreenSelectProfile and nothing would happen.

I'd never looked into it previously because of my own understanding that ScreenSelectProfile doesn't make sense on a StepMania machine where the MenuTimer is enabled (presumably a public arcade machine).  There's probably a very specific scenario in where it's wanted and the community using the machine is stable and everyone trusts one another, but USB profiles generally make more sense for public machines.

Anyway, this commit fixes the MenuTimer to force a screen transition when it reaches 0 on SelectProfile.

If both players are joined, and have the same profile selected, and that profile isn't GUEST, I opted to set both players to GUEST rather than setting both players to the same local profile.