perform additional validation on player profiles
Commit 296fd14adc86dbc0767794f46760212765e472ed switched SL's internal use of lowercase "x" to uppercase "X" when referring to SpeedModType.  This allowed me to replace some verbose string-comparison-based branching (if "x" then ... elseif "C" then ... elseif "M" then ...) and it is intended to help a larger restructuring of ScreenPlayerOptions that is currently running ~10 days behind schedule.

Switching from "x" to "X" internally also introduced a new bug.  Player profiles that had "x" saved as the value for "SpeedModType" from previous games would break SL.  It would pass the existing validation ("SpeedModType" is a valid key to save to profile, and "x" is a string) but "x" was no longer a valid choice in the OptionRow for SpeedModType.

Thus, this commit adds additional validation to ensure that string settings saved to profile are actually valid choices in the corresponding OptionRow.

If an invalid string is specified (for example, SpeedModType=nice in a player's local profile) the fallback value from ./Scripts/SL_Init.lua will be used and saved to profile the next time SaveProfileCustom() is called.

This commit has prrrrrrrrobably introduced more bugs that I haven't yet detected.